# backend/client_openapi.yaml
openapi: 3.0.0
info:
  title: Snake Game Client API
  version: 1.0.0
  description: API requirements based on frontend/src/api/mockClient.ts

paths:
  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
                  error:
                    type: string

  /auth/signup:
    post:
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Signup success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
                  error:
                    type: string

  /leaderboard:
    get:
      summary: Get leaderboard entries
      responses:
        '200':
          description: List of top scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'

  /games/active:
    get:
      summary: Get currently active games
      responses:
        '200':
          description: List of active games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActiveGame'

  /scores:
    post:
      summary: Update high score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - score
              properties:
                userId:
                  type: string
                score:
                  type: integer
      responses:
        '200':
          description: Score updated successfully

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        avatar:
          type: string
        highScore:
          type: integer

    LeaderboardEntry:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        avatar:
          type: string
        score:
          type: integer
        date:
          type: string
          format: date-time

    ActiveGame:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        avatar:
          type: string
        score:
          type: integer
        gameMode:
          type: string
          enum: [classic, portal]